+++ src/main/java/com/guno/dataimport/mapper/ShopeeMapper.java
@@ -605,9 +605,9 @@ public class ShopeeMapper {
         return null;
     }
 
-    private Integer getCurrentDateKey() {
+    private Long getCurrentDateKey() {
         LocalDateTime now = LocalDateTime.now();
-        return Integer.parseInt(now.format(DateTimeFormatter.ofPattern("yyyyMMdd")));
+        return Long.parseLong(now.format(DateTimeFormatter.ofPattern("yyyyMMdd")));
     }
 
     private Long generateDateKey(LocalDateTime dateTime) {
